Changed the way it works
The plugin can be called as soon as possible to initialize it. It will
wait for the ready event to show the banner. Added simple options to
make it easier to use in simple setups. Added disable option so it stays
shut when needed. README updated and reformatted a bit. Demo updated and
made simpler.